requirements.txt
# DevSquad Dependencies

# Core
pyyaml>=6.0
python>=3.9

# Standard library modules used (no install needed):
# - concurrent.futures (parallel execution)
# - dataclasses (data models)
# - json (data processing)
# - pathlib (path handling)
# - typing (type annotations)
# - logging (logging)
# - argparse (CLI arguments)
# - re (regex patterns)
# - hashlib (SHA256 integrity)
# - ast (code analysis)
# - datetime (timestamps)
# - sqlite3 (database)
# - tempfile (temp directories)
# - yaml (optional — for config file support)

# Optional LLM Backends (install only what you need)
# openai>=1.0          # OpenAI backend (pip install devsquad[openai])
# anthropic>=0.18       # Anthropic backend (pip install devsquad[anthropic])

# Optional Memory Layer
# carrymem>=0.8        # CarryMem portable AI memory (pip install devsquad[carrymem])

# Optional MCP Server
# mcp>=0.9             # MCP protocol (pip install devsquad[mcp])

# Development (pip install devsquad[dev])
# pytest>=7.0
# pytest-asyncio>=0.21
# black>=23.0
# flake8>=6.0
# mypy>=1.0

# Install all optional: pip install devsquad[openai,anthropic,mcp,dev]
